+++ hephaistos/config.py
@@ -48,3 +48,4 @@ scale_factor_X: float
 scale_factor_Y: float
 scale_factor: float
 center_hud = False
+modimporter: Path = None